backend/pyproject.toml
[tool.poetry]
name = "trackedfitness-backend"
version = "0.1.0"
description = "Fitness tracking web application."
authors = ["Michael Jarvis <mjarvis9541@gmail.com>"]
readme = "README.md"
packages = [{include = "trackedfitness_backend"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.14.0"
python-dotenv = "^1.0.0"
django-filter = "^22.1"
pyjwt = "^2.6.0"
pillow = "^9.4.0"
gunicorn = "^20.1.0"
factory-boy = "^3.2.1"
django-storages = "^1.13.2"
boto3 = "^1.26.84"
psycopg2 = "^2.9.6"


[tool.poetry.group.dev.dependencies]
django-extensions = "^3.2.1"
django-debug-toolbar = "^3.8.1"
ipython = "^8.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
